TRIPLEX-830 Доработки сторей
