+++ stories/Badge/examples/SizesExample.tsx
@@ -14,7 +14,7 @@ import {
 export const SizesExample = () => (
     <div style={{ display: "flex", flexDirection: "column", gap: "16px", maxWidth: "150px" }}>
         <div style={{ display: "flex", flexDirection: "column", gap: "8px" }}>
-            <div style={{ marginBottom: "8px", fontSize: "16px", fontWeight: "700" }}>EComponentSize.SM</div>
+            <div style={{ marginBottom: "8px", fontSize: "16px", fontWeight: "700" }}>SM</div>
             <Badge size={EComponentSize.SM}>
                 <Caption size={ECaptionSize.C1} weight={EFontWeightCaption.REGULAR} type={EFontType.PRIMARY_INVERT}>
                     Badge text SM
@@ -22,7 +22,7 @@ export const SizesExample = () => (
             </Badge>
         </div>
         <div style={{ display: "flex", flexDirection: "column", gap: "8px" }}>
-            <div style={{ marginBottom: "8px", fontSize: "16px", fontWeight: "700" }}>EComponentSize.MD</div>
+            <div style={{ marginBottom: "8px", fontSize: "16px", fontWeight: "700" }}>MD</div>
             <Badge size={EComponentSize.MD}>
                 <Text size={ETextSize.B4} weight={EFontWeightText.REGULAR} type={EFontType.PRIMARY_INVERT}>
                     Badge text MD
@@ -30,7 +30,7 @@ export const SizesExample = () => (
             </Badge>
         </div>
         <div style={{ display: "flex", flexDirection: "column", gap: "8px" }}>
-            <div style={{ marginBottom: "8px", fontSize: "16px", fontWeight: "700" }}>EComponentSize.LG</div>
+            <div style={{ marginBottom: "8px", fontSize: "16px", fontWeight: "700" }}>LG</div>
             <Badge size={EComponentSize.LG}>
                 <Text size={ETextSize.B3} weight={EFontWeightText.REGULAR} type={EFontType.PRIMARY_INVERT}>
                     Badge text LG
