+++ stories/Chips/ChipSort.stories.tsx
@@ -1,11 +1,10 @@
 import React, { useEffect, useRef, useState } from "react";
-import { Meta, StoryObj } from "@storybook/react";
+import { StoryObj } from "@storybook/react";
+import { ChipSort } from "../../src/components/Chip/ChipSort";
+import { EComponentSize } from "../../src/enums/EComponentSize";
 import { Title, Description, Primary, Controls, Stories, Heading, ArgTypes } from "@storybook/addon-docs/blocks";
-import { ChipSort, EComponentSize } from "@sberbusiness/triplex-next";
-import { ISelectFieldOption } from "../../src/components/SelectField/SelectField";
-import { DefaultExample, DefaultExampleSource, SizesExample, SizesExampleSource } from "./examples/ChipSort";
 
-const meta = {
+export default {
     title: "Components/Chips/ChipSort",
     component: ChipSort,
     tags: ["autodocs"],
@@ -30,17 +29,9 @@ ChipSelect с иконкой выбора сортировки.
             ),
         },
     },
-} satisfies Meta<typeof ChipSort>;
+};
 
-export default meta;
-type Story = StoryObj<typeof ChipSort>;
-type PlaygroundStory = StoryObj<
-    Omit<React.ComponentProps<typeof ChipSort>, "defaultValue"> & {
-        defaultValue: number;
-    }
->;
-
-export const Playground: PlaygroundStory = {
+export const Playground: StoryObj<typeof ChipSort> = {
     tags: ["!autodocs"],
     args: {
         size: EComponentSize.MD,
@@ -87,24 +78,67 @@ export const Playground: PlaygroundStory = {
     },
 };
 
-export const Default: Story = {
-    render: DefaultExample,
+export const Default: StoryObj<typeof ChipSort> = {
     parameters: {
         controls: { disable: true },
-        docs: { source: { code: DefaultExampleSource, language: "tsx" } },
+    },
+    render: () => {
+        const options = [
+            { id: "chip-sort-1", label: "По дате", value: "i1" },
+            { id: "chip-sort-2", label: "По времени", value: "i2" },
+            { id: "chip-sort-3", label: "По названию", value: "i3" },
+        ];
+
+        const [value, setValue] = useState(options[0]);
+
+        return <ChipSort defaultValue={options[0]} value={value} options={options} onChange={setValue} />;
     },
 };
 
-export const Sizes: Story = {
-    render: SizesExample,
+export const Sizes: StoryObj<typeof ChipSort> = {
     parameters: {
         controls: { disable: true },
-        docs: { source: { code: SizesExampleSource, language: "tsx" } },
+    },
+    render: () => {
+        const options = [
+            { id: "chip-sort-1", label: "По дате", value: "i1" },
+            { id: "chip-sort-2", label: "По времени", value: "i2" },
+            { id: "chip-sort-3", label: "По названию", value: "i3" },
+        ];
+
+        const [valueSM, setValueSM] = useState(options[0]);
+        const [valueMD, setValueMD] = useState(options[0]);
+        const [valueLG, setValueLG] = useState(options[0]);
+
+        return (
+            <div style={{ display: "flex", flexDirection: "column", alignItems: "flex-start", gap: "20px" }}>
+                <ChipSort
+                    defaultValue={options[0]}
+                    size={EComponentSize.SM}
+                    value={valueSM}
+                    options={options}
+                    onChange={setValueSM}
+                />
+                <ChipSort
+                    defaultValue={options[0]}
+                    size={EComponentSize.MD}
+                    value={valueMD}
+                    options={options}
+                    onChange={setValueMD}
+                />
+                <ChipSort
+                    defaultValue={options[0]}
+                    size={EComponentSize.LG}
+                    value={valueLG}
+                    options={options}
+                    onChange={setValueLG}
+                />
+            </div>
+        );
     },
 };
 
-export const VisualTests: Story = {
-    tags: ["!autodocs"],
+export const VisualTests: StoryObj<typeof ChipSort> = {
     parameters: {
         controls: { disable: true },
         docs: {
@@ -115,14 +149,16 @@ export const VisualTests: Story = {
         },
     },
     render: () => {
-        const options: ISelectFieldOption[] = [
+        const options = [
             { id: "chip-sort-1", label: "По дате", value: "i1" },
             { id: "chip-sort-2", label: "По времени", value: "i2" },
             { id: "chip-sort-3", label: "По названию", value: "i3" },
         ];
-        const [valueSM, setValueSM] = useState<ISelectFieldOption>(options[0]);
-        const [valueMD, setValueMD] = useState<ISelectFieldOption>(options[0]);
-        const [valueLG, setValueLG] = useState<ISelectFieldOption>(options[0]);
+
+        const [valueSM, setValueSM] = useState(options[0]);
+        const [valueMD, setValueMD] = useState(options[0]);
+        const [valueLG, setValueLG] = useState(options[0]);
+
         const rootRef = useRef<HTMLDivElement>(null);
 
         useEffect(() => {
