+++ stories/Chips/ChipSuggest.stories.tsx
@@ -4,14 +4,6 @@ import { ChipSuggest } from "../../src/components/Chip/ChipSuggest/ChipSuggest";
 import { ISuggestFieldOption } from "../../src/components/SuggestField/types";
 import { EComponentSize } from "../../src/enums";
 import { Title, Description, Controls, Stories, Primary, Heading, ArgTypes } from "@storybook/addon-docs/blocks";
-import {
-    DefaultExample,
-    DefaultExampleSource,
-    SizesExample,
-    SizesExampleSource,
-    StatesExample,
-    StatesExampleSource,
-} from "./examples/ChipSuggest";
 
 const meta = {
     title: "Components/Chips/ChipSuggest",
@@ -139,6 +131,7 @@ export const Playground: StoryObj<typeof ChipSuggest> = {
             "Olive",
             "Pea Pod",
         ];
+
         const initialOptions: ISuggestFieldOption[] = fruits.map((fruit, index) => ({
             id: `suggest-option-${index}`,
             label: fruit,
@@ -168,6 +161,7 @@ export const Playground: StoryObj<typeof ChipSuggest> = {
             setOptions(filteredOptions);
             setTooltipOpen(filteredOptions.length === 0);
         };
+
         return (
             <ChipSuggest
                 {...args}
@@ -184,31 +178,297 @@ export const Playground: StoryObj<typeof ChipSuggest> = {
 };
 
 export const Default: StoryObj<typeof ChipSuggest> = {
-    render: DefaultExample,
     parameters: {
         controls: { disable: true },
-        docs: { source: { code: DefaultExampleSource, language: "tsx" } },
+    },
+    render: () => {
+        const fruits = [
+            "Hot Pepper",
+            "Corn",
+            "Tomato",
+            "Eggplant",
+            "Grapes",
+            "Melon",
+            "Watermelon",
+            "Tangerine",
+            "Lemon",
+            "Banana",
+            "Pineapple",
+            "Red Apple",
+            "Green Apple",
+            "Pear",
+            "Peach",
+            "Cherries",
+            "Strawberry",
+            "Avocado",
+            "Cucumber",
+            "Kiwi",
+            "Coconut",
+            "Mango",
+            "Blueberries",
+            "Bell Pepper",
+            "Olive",
+            "Pea Pod",
+        ];
+
+        const initialOptions: ISuggestFieldOption[] = fruits.map((fruit, index) => ({
+            id: `suggest-option-${index}`,
+            label: fruit,
+        }));
+
+        const [value, setValue] = useState<ISuggestFieldOption>();
+        const [options, setOptions] = useState<ISuggestFieldOption[]>([]);
+        const [tooltipOpen, setTooltipOpen] = useState(false);
+        const initialOptionsRef = useRef<ISuggestFieldOption[]>(initialOptions);
+
+        const handleDropdownOpen = () => {
+            setOptions(initialOptionsRef.current);
+            setTooltipOpen(false);
+        };
+
+        const handleFilter = (inputValue: string) => {
+            if (inputValue.length === 0) {
+                setOptions(initialOptionsRef.current);
+                setTooltipOpen(false);
+                return;
+            }
+
+            const filteredOptions = initialOptionsRef.current.filter(({ label }) =>
+                label.toLowerCase().includes(inputValue.toLowerCase()),
+            );
+
+            setOptions(filteredOptions);
+            setTooltipOpen(filteredOptions.length === 0);
+        };
+
+        return (
+            <ChipSuggest
+                value={value}
+                options={options}
+                tooltipOpen={tooltipOpen}
+                onSelect={setValue}
+                onFilter={handleFilter}
+                targetProps={{ clearSelected: () => setValue(undefined) }}
+                dropdownProps={{ onOpen: handleDropdownOpen }}
+                label="Suggest label"
+                placeholder="Type to proceed"
+                noOptionsText="No matches found."
+                size={EComponentSize.MD}
+            />
+        );
     },
 };
 
 export const Sizes: StoryObj<typeof ChipSuggest> = {
-    render: SizesExample,
     parameters: {
         controls: { disable: true },
-        docs: { source: { code: SizesExampleSource, language: "tsx" } },
+    },
+    render: () => {
+        const sizes = Object.values(EComponentSize);
+
+        const fruits = [
+            "Hot Pepper",
+            "Corn",
+            "Tomato",
+            "Eggplant",
+            "Grapes",
+            "Melon",
+            "Watermelon",
+            "Tangerine",
+            "Lemon",
+            "Banana",
+            "Pineapple",
+            "Red Apple",
+            "Green Apple",
+            "Pear",
+            "Peach",
+            "Cherries",
+            "Strawberry",
+            "Avocado",
+            "Cucumber",
+            "Kiwi",
+            "Coconut",
+            "Mango",
+            "Blueberries",
+            "Bell Pepper",
+            "Olive",
+            "Pea Pod",
+        ];
+
+        const initialOptions: ISuggestFieldOption[] = fruits.map((fruit, index) => ({
+            id: `suggest-option-${index}`,
+            label: fruit,
+        }));
+
+        const [valueSM, setValueSM] = useState<ISuggestFieldOption>();
+        const [valueMD, setValueMD] = useState<ISuggestFieldOption>();
+        const [valueLG, setValueLG] = useState<ISuggestFieldOption>();
+
+        const [options, setOptions] = useState<ISuggestFieldOption[]>([]);
+        const [tooltipOpen, setTooltipOpen] = useState(false);
+        const initialOptionsRef = useRef<ISuggestFieldOption[]>(initialOptions);
+
+        const handleDropdownOpen = () => {
+            setOptions(initialOptionsRef.current);
+            setTooltipOpen(false);
+        };
+
+        const handleFilter = (inputValue: string) => {
+            if (inputValue.length === 0) {
+                setOptions(initialOptionsRef.current);
+                setTooltipOpen(false);
+                return;
+            }
+
+            const filteredOptions = initialOptionsRef.current.filter(({ label }) =>
+                label.toLowerCase().includes(inputValue.toLowerCase()),
+            );
+
+            setOptions(filteredOptions);
+            setTooltipOpen(filteredOptions.length === 0);
+        };
+
+        return (
+            <div style={{ display: "flex", flexDirection: "column", alignItems: "flex-start", gap: "20px" }}>
+                {sizes.map((size) => {
+                    return (
+                        <ChipSuggest
+                            key={size}
+                            size={size}
+                            label={size.toUpperCase()}
+                            placeholder="Type to proceed"
+                            noOptionsText="No matches found."
+                            value={
+                                size === EComponentSize.SM ? valueSM : size === EComponentSize.MD ? valueMD : valueLG
+                            }
+                            options={options}
+                            tooltipOpen={tooltipOpen}
+                            onSelect={
+                                size === EComponentSize.SM
+                                    ? setValueSM
+                                    : size === EComponentSize.MD
+                                      ? setValueMD
+                                      : setValueLG
+                            }
+                            onFilter={handleFilter}
+                            targetProps={{
+                                clearSelected: () =>
+                                    size === EComponentSize.SM
+                                        ? setValueSM(undefined)
+                                        : size === EComponentSize.MD
+                                          ? setValueMD(undefined)
+                                          : setValueLG(undefined),
+                            }}
+                            dropdownProps={{ onOpen: handleDropdownOpen }}
+                        />
+                    );
+                })}
+            </div>
+        );
     },
 };
 
 export const States: StoryObj<typeof ChipSuggest> = {
-    render: StatesExample,
     parameters: {
         controls: { disable: true },
         docs: {
             description: {
                 story: "ChipSuggest в состояниях selected, disabled.",
             },
-            source: { code: StatesExampleSource, language: "tsx" },
         },
+    },
+    render: () => {
+        const fruits = [
+            "Hot Pepper",
+            "Corn",
+            "Tomato",
+            "Eggplant",
+            "Grapes",
+            "Melon",
+            "Watermelon",
+            "Tangerine",
+            "Lemon",
+            "Banana",
+            "Pineapple",
+            "Red Apple",
+            "Green Apple",
+            "Pear",
+            "Peach",
+            "Cherries",
+            "Strawberry",
+            "Avocado",
+            "Cucumber",
+            "Kiwi",
+            "Coconut",
+            "Mango",
+            "Blueberries",
+            "Bell Pepper",
+            "Olive",
+            "Pea Pod",
+        ];
+
+        const initialOptions: ISuggestFieldOption[] = fruits.map((fruit, index) => ({
+            id: `suggest-option-${index}`,
+            label: fruit,
+        }));
+
+        const [valueSelected, setValueSelected] = useState<ISuggestFieldOption>();
+        const [valueDisabled, setValueDisabled] = useState<ISuggestFieldOption>();
+        const [options, setOptions] = useState<ISuggestFieldOption[]>([]);
+        const [tooltipOpen, setTooltipOpen] = useState(false);
+        const initialOptionsRef = useRef<ISuggestFieldOption[]>(initialOptions);
+
+        const handleDropdownOpen = () => {
+            setOptions(initialOptionsRef.current);
+            setTooltipOpen(false);
+        };
+
+        const handleFilter = (inputValue: string) => {
+            if (inputValue.length === 0) {
+                setOptions(initialOptionsRef.current);
+                setTooltipOpen(false);
+                return;
+            }
+
+            const filteredOptions = initialOptionsRef.current.filter(({ label }) =>
+                label.toLowerCase().includes(inputValue.toLowerCase()),
+            );
+
+            setOptions(filteredOptions);
+            setTooltipOpen(filteredOptions.length === 0);
+        };
+
+        return (
+            <div style={{ display: "flex", flexDirection: "column", alignItems: "flex-start", gap: "20px" }}>
+                {[
+                    { label: "Suggest label", targetProps: { selected: true } },
+                    { label: "Suggest label", targetProps: { disabled: true } },
+                ].map((state) => {
+                    return (
+                        <ChipSuggest
+                            key={state.label}
+                            label={state.targetProps.selected ? "Selected" : "Disabled"}
+                            size={EComponentSize.MD}
+                            placeholder="Type to proceed"
+                            noOptionsText="No matches found."
+                            value={state.targetProps.selected ? valueSelected : valueDisabled}
+                            options={options}
+                            tooltipOpen={tooltipOpen}
+                            onSelect={state.targetProps.selected ? setValueSelected : setValueDisabled}
+                            onFilter={handleFilter}
+                            targetProps={{
+                                clearSelected: () =>
+                                    state.targetProps.selected
+                                        ? setValueSelected(undefined)
+                                        : setValueDisabled(undefined),
+                                ...state.targetProps,
+                            }}
+                            dropdownProps={{ onOpen: handleDropdownOpen }}
+                        />
+                    );
+                })}
+            </div>
+        );
     },
 };
 
@@ -275,7 +535,6 @@ export const WithNotificationIcon: StoryObj<typeof ChipSuggest> = {
 };
 
 export const VisualTests: StoryObj<typeof ChipSuggest> = {
-    tags: ["!autodocs"],
     parameters: {
         controls: { disable: true },
         docs: {
