--- stories/Chips/examples/ChipIcon/index.ts
@@ -1,4 +1,0 @@
-export { ChipIconDefaultExample as DefaultExample } from "./ChipIconDefaultExample";
-export { default as DefaultExampleSource } from "./ChipIconDefaultExample?raw";
-export { ChipIconSizesExample as SizesExample } from "./ChipIconSizesExample";
-export { default as SizesExampleSource } from "./ChipIconSizesExample?raw";
