--- stories/Chips/examples/ChipOptions/index.ts
@@ -1,4 +1,0 @@
-export { ChipOptionsDefaultExample as DefaultExample } from "./ChipOptionsDefaultExample";
-export { default as DefaultExampleSource } from "./ChipOptionsDefaultExample?raw";
-export { ChipOptionsSizesExample as SizesExample } from "./ChipOptionsSizesExample";
-export { default as SizesExampleSource } from "./ChipOptionsSizesExample?raw";
