--- stories/Chips/examples/ChipSelect/index.ts
@@ -1,8 +1,0 @@
-export { ChipSelectDefaultExample as DefaultExample } from "./ChipSelectDefaultExample";
-export { default as DefaultExampleSource } from "./ChipSelectDefaultExample?raw";
-export { ChipSelectSizesExample as SizesExample } from "./ChipSelectSizesExample";
-export { default as SizesExampleSource } from "./ChipSelectSizesExample?raw";
-export { ChipSelectStatesExample as StatesExample } from "./ChipSelectStatesExample";
-export { default as StatesExampleSource } from "./ChipSelectStatesExample?raw";
-export { ChipSelectWithCustomDisplayedValueExample as WithCustomDisplayedValueExample } from "./ChipSelectWithCustomDisplayedValueExample";
-export { default as WithCustomDisplayedValueExampleSource } from "./ChipSelectWithCustomDisplayedValueExample?raw";
