--- stories/Chips/examples/ChipSort/index.ts
@@ -1,4 +1,0 @@
-export { ChipSortDefaultExample as DefaultExample } from "./ChipSortDefaultExample";
-export { default as DefaultExampleSource } from "./ChipSortDefaultExample?raw";
-export { ChipSortSizesExample as SizesExample } from "./ChipSortSizesExample";
-export { default as SizesExampleSource } from "./ChipSortSizesExample?raw";
