--- stories/Chips/examples/ChipSuggest/index.ts
@@ -1,6 +1,0 @@
-export { ChipSuggestDefaultExample as DefaultExample } from "./ChipSuggestDefaultExample";
-export { default as DefaultExampleSource } from "./ChipSuggestDefaultExample?raw";
-export { ChipSuggestSizesExample as SizesExample } from "./ChipSuggestSizesExample";
-export { default as SizesExampleSource } from "./ChipSuggestSizesExample?raw";
-export { ChipSuggestStatesExample as StatesExample } from "./ChipSuggestStatesExample";
-export { default as StatesExampleSource } from "./ChipSuggestStatesExample?raw";
